fix(search-api): address 25 review findings across Search API types
Critical fixes:
- Add else-throw in buildKnnRankMap for unrecognized query types (#1)
- Fix groups() to return empty list when not grouped, matching interface contract (#2)
- Change SearchResultRow.getScore() from Float to Double for wire precision (#3)
- Add defensive copy in Knn.getQuery() for float[] queries (#4)
- Fix Knn.queryText() Javadoc: text sent to server, not client-side embedded (#5)

Important fixes:
- Add null dto check in SearchResultImpl.from() (#6)
- Add bounds validation with actionable message in rows(searchIndex) (#7)
- Add Objects.requireNonNull to Search.Builder setters (knn, rrf, where, groupBy, select) (#8)
- Add null validation to Knn.key() (#9)
- Validate null elements in SearchBuilderImpl.searches() (#10)
- Rename groupCount() to searchCount() with clarified Javadoc (#13)

Suggestions:
- Add minK/maxK range and ordering validation in GroupBy.build() (#14)
- Make Rrf.RankWithWeight fields private (#15)
- Add null rows check in SearchResultGroupImpl (#16)
- Add defensive rank-missing check in buildSearchItemMap (#18)
- Remove internal "D-04" reference from production comment (#23)
- Make ResultRow Javadoc mechanism-agnostic (Include/Select) (#24)
- Clarify SearchBuilder limit/offset as per-search fallback (#25)
- Add mutual-exclusivity docs to SearchBuilder convenience methods
- Add IllegalArgumentException to SearchBuilder.execute() @throws

+++ src/main/java/tech/amikos/chromadb/v2/SearchResultRow.java
@@ -12,7 +12,8 @@ public interface SearchResultRow extends ResultRow {
      * Returns the relevance score from the ranking expression, or {@code null} if
      * {@link Select#SCORE} was not included in the projection.
      *
-     * <p>Higher scores indicate greater relevance.</p>
+     * <p>Higher scores indicate greater relevance. Returns {@link Double} to preserve
+     * the full wire-format precision, consistent with {@link SearchResult#getScores()}.</p>
      */
-    Float getScore();
+    Double getScore();
 }
